simplify tests and make it 2
