+++ test/fixtures.ts
@@ -37,6 +37,7 @@ import {
   commitEpoch,
   getCorrectVotingTimestamp,
 } from "./helper";
+import { expect } from "chai";
 
 // --------------- Deploying Contracts ---------------
 
@@ -359,6 +360,64 @@ async function initializedWithSpecificBonusesStateFixtureFunction(this: Mocha.Co
     value: this.minStake.mul(500),
   });
 
+  // Check if the bonuses are set correctly
+  expect(await aprCalculator.getBaseAPR()).to.be.equal(500);
+  expect(await aprCalculator.getMacroFactor()).to.be.equal(2500);
+  expect(await aprCalculator.getRSIBonus()).to.be.equal(11500);
+
+  // Disable whitelisting
+  await hydraChain.connect(this.signers.governance).disableWhitelisting();
+  expect(await hydraChain.isWhitelistingEnabled()).to.be.false;
+
+  // Commit epoch
+  const epochId = hre.ethers.BigNumber.from(1);
+  const epoch = {
+    startBlock: hre.ethers.BigNumber.from(1),
+    endBlock: hre.ethers.BigNumber.from(500),
+    epochRoot: this.epoch.epochRoot,
+  };
+  const uptime = [
+    {
+      validator: this.signers.admin.address,
+      signedBlocks: hre.ethers.BigNumber.from(10),
+    },
+  ];
+  await systemHydraChain.commitEpoch(epochId, epoch, 500, uptime);
+  await hydraStaking.connect(this.signers.system).distributeRewardsFor(epochId, uptime);
+
+  // register validators
+  const keyPair = mcl.newKeyPair();
+  const validator1signature = mcl.signValidatorMessage(
+    DOMAIN,
+    CHAIN_ID,
+    this.signers.validators[0].address,
+    keyPair.secret
+  ).signature;
+
+  const validator2signature = mcl.signValidatorMessage(
+    DOMAIN,
+    CHAIN_ID,
+    this.signers.validators[1].address,
+    keyPair.secret
+  ).signature;
+
+  const validator3signature = mcl.signValidatorMessage(
+    DOMAIN,
+    CHAIN_ID,
+    this.signers.validators[2].address,
+    keyPair.secret
+  ).signature;
+
+  await hydraChain
+    .connect(this.signers.validators[0])
+    .register(mcl.g1ToHex(validator1signature), mcl.g2ToHex(keyPair.pubkey));
+  await hydraChain
+    .connect(this.signers.validators[1])
+    .register(mcl.g1ToHex(validator2signature), mcl.g2ToHex(keyPair.pubkey));
+  await hydraChain
+    .connect(this.signers.validators[2])
+    .register(mcl.g1ToHex(validator3signature), mcl.g2ToHex(keyPair.pubkey));
+
   return {
     hydraChain,
     systemHydraChain,
@@ -372,6 +431,9 @@ async function initializedWithSpecificBonusesStateFixtureFunction(this: Mocha.Co
     priceOracle,
     rewardWallet,
     DAOIncentiveVault,
+    validator1: this.signers.validators[0],
+    validator2: this.signers.validators[1],
+    validator3: this.signers.validators[2],
   };
 }
